backend/requirements.txt
Package              Version
-------------------- ---------
asgiref              3.5.2
certifi              2022.9.24
Django               4.1.2
django-debug-toolbar 3.7.0
djangorestframework  3.14.0
pip                  22.2.2
pytz                 2022.5
setuptools           63.4.1
sqlparse             0.4.3
wheel                0.37.1
